readRelayQueryData: set metadata when creating fragment pointers
Summary: Fixes an issue that occurs when reading the results of a field whose only child is a fragment reference. D2947985 changed the reader to set fragment pointers on the result object directly, bypassing the call to `_setDataValue`. That bypassed the method's side-effect of also writing all the metadata fields on the object. The fix here is to explicitly call a new `_setMetadataFields` method to ensure that those fields are set (see the new test).

Reviewed By: yuzhi

Differential Revision: D2959408

fb-gh-sync-id: 8dacfe4f6f1dd44ed1176299d2f03620f2c77ec7
shipit-source-id: 8dacfe4f6f1dd44ed1176299d2f03620f2c77ec7