Use known typename for existing records
Summary:Relay doesn't include the `__typename` field on queries/fragments/fields that have a concrete type. This means that there can be times (see issue #474) where the query/payload legitimately doesn't have `__typename`, but the writer attempts to read it, sees that it isn't there, and warns. In these cases we should instead check to see if we already know the type of the record. Doing so fixes the false-positive "missing typename" warning.

Closes https://github.com/facebook/relay/pull/860

Reviewed By: yungsters

Differential Revision: D2958749

fb-gh-sync-id: 781ebf3b89f17fa94bb8d823e9139c39d69fabe6
shipit-source-id: 781ebf3b89f17fa94bb8d823e9139c39d69fabe6