Fix query constructions from paths in OSS (thin interfaces)
Summary:Fix for #861. Copying my description there:

In this case, what's happening is that you're fetching some record:

```
query {
  viewer { # concrete type, id '123'
    # ... fields
    session { # concrete type, no id
      # ... fields
     }
  }
}
```

Because everything is a concrete type it's okay to query fields directly w/o conditioning on type (i.e. you don't need to do `viewer { ... on SomeType { session } }`). However when Relay tries to refetch `session`, it doesn't have an id, so it looks up the "path" to session from the nearest node that does have an id, in this case the `viewer` record. It constructs a query:

```
node(id: "123") { ... }
```

And fills in everything that was on the path from node 123 down to session:

```
node(id: "123") {
  session {
    # ...
  }
}
```

But this is invalid: while in the original query `session` was a direct field of a concrete type (so no need to condition), now it's a direct field on a different type, `Node`, which doesn't have a `session` field. Relay should use its knowledge of the type of record 123 to construct the conditioning fragment and generate the correct:

```
node(id: "123") {
  ... on Viewer {
    session {
      # ...
    }
  }
}
```

Closes https://github.com/facebook/relay/pull/862

Reviewed By: yungsters

Differential Revision: D2959873

fb-gh-sync-id: 0696c6d6da2efed7cc412586bbd7a0414f4f67f3
shipit-source-id: 0696c6d6da2efed7cc412586bbd7a0414f4f67f3