Relay: Fix Reading Empty Records
Summary:When we read connections or fragments, we don't actually allocate a record until we encounter a child field that is not null.

This revision fixes the bug by ensuring that we allocate the record for connections and fragments (if it exists) regardless of whether or not we encounter a child field that is not null.

Reviewed By: josephsavona

Differential Revision: D2998213

fb-gh-sync-id: 7d831a20e90b9b456ca2599102ea24cc0600d529
shipit-source-id: 7d831a20e90b9b456ca2599102ea24cc0600d529