babel plugin: support input object literals
Summary:This is an expanded portion of just the plugin changes from [#844](https://github.com/facebook/relay/pull/844). The goal of that PR is to support arbitrary values in identifying arguments. Part of that means supporting input objects, which we didn't fully support in the plugin (we only allowed the value of an input object argument to be a variable, not a literal).

Changes:
- Support parsing and printing literal InputObject expressions.
- Clear error message when variables appear nested in ways that the Relay runtime cannot interpret.
- Tests for the above.

Closes https://github.com/facebook/relay/pull/894

Reviewed By: yungsters

Differential Revision: D3002163

Pulled By: josephsavona

fb-gh-sync-id: d7708b36bc087bdd744c4d027cf4b888e8350798
shipit-source-id: d7708b36bc087bdd744c4d027cf4b888e8350798