babel plugin: always annotate argument types (including scalars)
Summary:Fix for #916 - `diffRelayQuery` is generating invalid plural root calls. The current logic assumes that if the identifying argument type is not present (because it is scalar) that the scalar must be `ID`. This isn't always the case, so the simplest option is to *always* print the type of root field identifying arguments.

Closes https://github.com/facebook/relay/pull/917

Reviewed By: steveluscher

Differential Revision: D3017121

Pulled By: josephsavona

fb-gh-sync-id: 9215599ef5f520ec13d0c5dc246d0dafb4c66b01
shipit-source-id: 9215599ef5f520ec13d0c5dc246d0dafb4c66b01