queued transaction can be also rollbacked
Summary:This is follow up to #607 that adds posibility to rollback transaction with status `COMMIT_QUEUED`.

There is one more missing piece for this use case that I was not sure how to approach.

To replace transaction that is queued I need to first check if its status is still `COMMIT_QUEUED`. Issue is that transaction is removed from pendingTransactionMap once its successfully committed. So I can easily end up with exception from `getStatus()`, which is bit overreaction :-).

So maybe return some status like `FINISHED` instead of throwing if its not in pending queue? I am of course very opened about naming :-).
Closes https://github.com/facebook/relay/pull/646

Reviewed By: josephsavona

Differential Revision: D3021199

Pulled By: yungsters

fb-gh-sync-id: 2ee1d2849d8bd5081ef57a42c136d644b61e95bd
shipit-source-id: 2ee1d2849d8bd5081ef57a42c136d644b61e95bd