Contextualize Relay mutations
Summary:== Original Description ==
This PR is another step toward making all Relay state contextual (#558). Originally this was submitted as a part of #683, but josephsavona [suggested](https://github.com/facebook/relay/pull/683#issuecomment-166386149) to split it to smaller PRs.

Currently, Relay resolve mutations props using the global `RelayStore`, and store them in the `props` instance field of `RelayMutation`. That makes it impossible to contextualize Relay mutations without changing the public API.

After analyzing different alternatives, I have found that the probably cleanest solution is not to store resolved mutation props in an instance field, but to pass them as an argument to each of the mutation methods. That way mutations can be reused in different `RelayStoreData` contexts, and also they can be reused in the same context but at different times, because props are re-resolved at each mutation execution.

== Summary of Changes ==
- Change `RelayMutation` constructor to not immediately resolve data for fragment props
- Add `bindContext` method on RelayMutation, called when the mutation is committed on a specific RelayContext
- This necessitated adding the `read` method to `RelayContextInterface`, and stubbing this method on alternative context implementations.
- Updated mutation docs to mention this method.

Closes https://github.com/facebook/relay/pull/699

Reviewed By: yungsters

Differential Revision: D2913334

Pulled By: josephsavona

fb-gh-sync-id: 5051fddf068fc8bc7379bb55c8b1e4d97e33313c
shipit-source-id: 5051fddf068fc8bc7379bb55c8b1e4d97e33313c