Treat all fat queries as patterns and preserve their empty nodes
Summary:So, here's the thing. Currently, the plugin allows you to write an invalid query: a subselection-supporting field that has no subselections:

    fragment on Viewer { todos }

That should fail, unless you've specifically marked the fragment as being a pattern:

    fragment on Viewer relay(pattern: true) {
      todos
    }

Pattern fragments are those on which certain validations are disabled.

For now, we're going to treat all fat queries as pattern fragments to unbreak products. Next, we're going to start to warn in the Babel Relay plugin when you have a query like the first one above. Lastly, we'll revert this PR and make the Babel Relay plugin throw on a query like the first one above.
Closes https://github.com/facebook/relay/pull/922

Reviewed By: wincent

Differential Revision: D3024309

Pulled By: steveluscher

fb-gh-sync-id: a1699e7a14149f215c0d97fb7141c3464e5ee67f
shipit-source-id: a1699e7a14149f215c0d97fb7141c3464e5ee67f