Relay: Recycle Relay Prop if Possible
Summary:If `RelayContainer.shouldComponentUpdate` returns true, we will attempt to recycle any existing resolved records (so that `prevProps` and `nextProps` are `===`). However, the `relay` prop is always recreated. This makes it so that using `RelayContainer` with any "pure" component (e.g. one that mixes in `RelayComponentWithPureRenderMixin`) cannot optimize when it compares the equality of `prevProps.relay` and `nextProps.relay`.

This revision changes `RelayContainer` to recycle the `relay` prop object whenever possible.

Reviewed By: steveluscher

Differential Revision: D3028906

fb-gh-sync-id: d46ed89e5eb23ecd4531701189642dd6bae83e66
shipit-source-id: d46ed89e5eb23ecd4531701189642dd6bae83e66