Relay: Create Mutation Hash on Record Props
Summary:After D3028906, `RelayContainer` no longer vends a new `relay` prop during every render. This change was made in order to allow components to use `ReactComponentWithPureRenderMixin`.

However, it became much easier to accidentally forget to check whether the status of any pending mutation transactions has changed. This is because the list of transactions and their status is hidden behind the invocation of `this.props.relay.getPendingTransactions()`.

In order to counteract this pit of failure, this revision adds a `__mutationHash__` onto vended records. The sole purpose of the `__mutationHash__` property is to ensure that a record, for which the sets of mutation transactions or their status has changed, will not be recycled. This means that using something like `ReactComponentWithPureRend` would continue to return update if no record data has changed but the mutation set or status have changed.

Reviewed By: josephsavona

Differential Revision: D3033377

fb-gh-sync-id: 546b932fdc33b17bede260aa56b038648b51b16c
shipit-source-id: 546b932fdc33b17bede260aa56b038648b51b16c