Relay: Reduce Scope of Tests to Relay Plugin
Summary:There is some broken integration between the Babel 6 plugins in `fbjs-scripts` and the transform tests in the `babel-relay-plugin`. For now, I am going to reduce the scope of the `babel-relay-plugin` tests until we're fully upgraded to Babel 6. This revision should fix the breaking tests in master.
Closes https://github.com/facebook/relay/pull/953

Reviewed By: josephsavona

Differential Revision: D3051944

Pulled By: yungsters

fb-gh-sync-id: b19643d2dbce17b6576b6f446fa0733328e8cb10
shipit-source-id: b19643d2dbce17b6576b6f446fa0733328e8cb10