Allow variable connection arguments
Summary:Relay Connections use a handful of arguments for pagination (first, last, before, after). These connection arguments are statically validated by the babel-relay-plugin, and then checked again at runtime by GraphQLRange. However, if the developer wants to use variables for the connection arguments (for example to support switching between `first` & `last` at runtime), the validation performed by babel-relay-plugin gets in the way and prevents this, which forces the developer to use some workarounds to "trick" the validation (see issue #677 for some examples).

This commit changes the babel-relay-plugin validation to allow otherwise invalid combinations (like `first`+`before`) as long as all the arguments are variables. When the connection arguments are variables, babel-relay-plugin effectively "trusts" that the developer will provide the correct arguments at runtime, and it becomes the job of GraphQLRange to validate the calls given the actual values. GraphQLRange was already doing this validation, so I simp
Closes https://github.com/facebook/relay/pull/708

Reviewed By: josephsavona

Differential Revision: D3042062

Pulled By: yungsters

fb-gh-sync-id: 6ee2409d7761a7ec58bf09f04fd2059f53a6d1fb
shipit-source-id: 6ee2409d7761a7ec58bf09f04fd2059f53a6d1fb