babel plugin: use fragment name as id when isStaticFragment=true
Summary:Now that we support `relay(isStaticFragment: true)` to denote static fragments (thanks yungsters!), we can use the fragment name as the id for static fragments instead of a hash. This means the identifier is stable across changes to the file, making it a bit easier to work with.
Closes https://github.com/facebook/relay/pull/943

Reviewed By: yungsters

Differential Revision: D3037308

Pulled By: josephsavona

fb-gh-sync-id: 922ce7afc5c49b39d76c5fea1c5f92616679ffd0
shipit-source-id: 922ce7afc5c49b39d76c5fea1c5f92616679ffd0