Relay: Ignore Retries w/o Failure
Summary:Currently, the `retry` functions supplied as a render argument by `RelayRenderer` throws if it is called when an error has not occurred. This was originally intended to discourage caching references to the function as a mechanism to trigger arbitrary fetch calls.

However, this invariant actually ends up being error prone in situations where users may accidentally trigger the invocation of the `retry` function in quick succession. The second invocation may occur before the request of the first has completed (and failed), thereby accidentally triggering the invariant violation.

This revision simply changes the invariant into a conditional so that subsequent invocations of the `retry` function do nothing.

Reviewed By: nickholub

Differential Revision: D3074256

fb-gh-sync-id: 13556c6430e22a209cdbbcccb06880ca87b0e3c9
shipit-source-id: 13556c6430e22a209cdbbcccb06880ca87b0e3c9