Simplify subscriptions to (possibly null) linked records
Summary:Relay uses subscriptions on record ids to inform Containers when their data may have changed. This logic is split between the reader, writer, and change emitter:
- the reader returns a list of "seen" record ids; the list of ids that, if changed in the store, could cause the results of the read operation to be different.
- the writer returns a list of ids that have changed as a result of applying the payload.
- "change" events are broadcast only on "updates" records - those that were already non-null and where a field changed - not on "created" records.

The one odd case is when a parent record links to a child record that has been deleted, removed from the store, or is not loaded yet due to incremental disk cache reads. In these cases the writer currently checks whether the linked record is "new" (i.e. it was created during the write) and if so records the parent record as having changed (even though nothing about the parent changed per-se). This is necessary because the reader would not have returned that child record as "seen" and isn't subscribed to that id.

This diff implements a simpler alternative: make sure the reader always includes linked ids as "seen" even if the record itself is null. The writer now only needs to record a parent as "changed" if the link itself actually changed (ie to point to a different child). In addition, it is necessary to broadcast updates on "created" records, since it is possible that the only record that could change within a fragment is that a link to a previously null record is now a link to an existent one.

The primary motivation behind this diff is GraphMode: the changes here allow records to be written in an arbitrary order and for "changes" to be recorded purely based on the previous/next value of fields without additional context (which isn't available in GraphMode since the linked record may not have been processed yet).

Reviewed By: yungsters

Differential Revision: D3070913

fb-gh-sync-id: e8296614141b7cfec0db92151763dc79c22dec12
shipit-source-id: e8296614141b7cfec0db92151763dc79c22dec12