Add `logNode` to `RelayTestUtils`
Summary:Sometimes while working on a test I want to just peek at some node which may or may not be printable. This commit adds `logNode`, which is a convenience wrapper around `flattenRelayQuery` + `printRelayQuery`, and which will wrap otherwise unprintable fields in a phony wrapper.

So, you can do `RelayTestUtils.logNode(node)` and it will print:

```
{
  "text": "fragment __PrintableFieldWrapper on __Phony{feedback{doesViewerLike,id}}",
  "variables": {}
}
```

Compare that with the harder to read result of `console.log(node)`:

```
RelayQueryField {
  __concreteNode__:
   { children: [ [Object], [Object], [Object] ],
     fieldName: 'feedback',
     kind: 'Field',
     metadata:
      { canHaveSubselections: true,
        inferredRootCallName: 'node',
        inferredPrimaryKey: 'id' },
     type: 'Feedback' },
  __route__: RelayMetaRoute { name: '$fromGraphQL' },
  __variables__: {},
  __calls__: [],
  __children__:
   [ RelayQueryField {
       __concreteNode__: [Object],
       _
Closes https://github.com/facebook/relay/pull/986

Reviewed By: josephsavona

Differential Revision: D3089702

Pulled By: wincent

fb-gh-sync-id: 849b8844033117bc30937d1a9bcc956942d76fe9
shipit-source-id: 849b8844033117bc30937d1a9bcc956942d76fe9