Relay: Revert `hasFragmentData` "Fix"
Summary:We cannot actually check if the supplied fragment reference in `hasFragmentData` is deferred because the instance of `RelayFragmentReference` does not actually have the "deferred" flag. No one ever supplies `getFragment('X').defer()` (and requiring this would defeat the purpose anyway).

For now, this check actually breaks `hasFragmentData` in legitimate use cases, so I'm reverting it.

Reviewed By: josephsavona

Differential Revision: D3104529

fb-gh-sync-id: ccaca2e187c87c4db6b5a27a75559ffd7f1dfed7
fbshipit-source-id: ccaca2e187c87c4db6b5a27a75559ffd7f1dfed7