Be more explicit about trackers
Summary:In these files we have change trackers, query trackers, and now fragment trackers. This can be confusing.

This commit makes sure that whenever we have more than one tracker in a local area, we use explicit names:

- `_tracker` becomes `_queryTracker`
- `tracker` becomes `queryTracker`

And so on. Even if we simplify things in the future to involve fewer trackers I still think this is a worthwhile change (another one is probably tackling the code smell of `writePayload` and the `RelayQueryWriter` constructor taking seven and six parameters respectively, but that is beyond the scope of this commit).
Closes https://github.com/facebook/relay/pull/1004

Reviewed By: josephsavona

Differential Revision: D3114338

Pulled By: wincent

fb-gh-sync-id: 1b44b078f281c60c67f5fb2c8a2c263aaa6c1832
fbshipit-source-id: 1b44b078f281c60c67f5fb2c8a2c263aaa6c1832