Relax connectionName validation slightly
In f2aa5b0e78c2f5 (D3095672) we tightened up mutation config validation a little too much, breaking product code that didn't explicitly include the connection field in the fat query definition.

In this commit we make it a little less aggressive:

- If the connection field is explicit in the fat query, we validate it.
- If it is not explicit but we can get it by intersecting with a tracked query, we validate that.
- Otherwise we let things slide and give up.

Note that in doing this I addressed the TODO comment and moved the code that deals with fields on the fat parent outside the conditional, so that it runs even when we don't have any tracked queries. That TODO predates even the open source release of Relay, so if somebody has context on if this is a terrible idea, please let me know.