Relax connectionName validation slightly
Summary:In f2aa5b0e78c2f5 (D3095672) we tightened up mutation config validation a little too much, breaking product code that didn't explicitly include the connection field in the fat query definition.

In this commit we make it a little less aggressive:

- If the connection field is explicit in the fat query, we validate it.
- If it is not explicit but we can get it by intersecting with a tracked query, we validate that.
- Otherwise we let things slide and give up.

Note that in doing this I addressed the TODO comment and moved the code that deals with fields on the fat parent outside the conditional, so that it runs even when we don't have any tracked queries. That TODO predates even the open source release of Relay, so if somebody has context on if this is a terrible idea, please let me know.
Closes https://github.com/facebook/relay/pull/1017

Reviewed By: yungsters

Differential Revision: D3136035

Pulled By: wincent

fb-gh-sync-id: d73d82d7cba647cace05d85e4c2f8919c6273490
fbshipit-source-id: d73d82d7cba647cace05d85e4c2f8919c6273490