RelayNodeInterface.getResultsFromPayload: don't return dataID
Summary: Preparation diff for GraphMode. The current writer traverses over query results with `RelayNodeInterface.getResultsFromPayload`; this method has the side effect of generating a new `dataID` for id-less root records. In the new GraphMode writer root ids aren't generated until a `putRoot` record is encountered and the dataID does not need to be generated in `getResutlsFromPayload`. This diff will reduce the changes when we actually switch on the GraphMode writer.

Reviewed By: wincent

Differential Revision: D3139655

fb-gh-sync-id: f4d579b24a23fda732e90063a965d09487b0fb20
fbshipit-source-id: f4d579b24a23fda732e90063a965d09487b0fb20