Update TodoMVC .babelrc for production builds of RN
Summary:It seems the presence of a `.babelrc` file in the react-native project is messing up the packager.
I'm not sure why but the bundling process was failing when hitting flow definitions.

Since the `.babelrc` configuration is needed when using relay with RN it can't be deleted.

I added an identical configuration for `production` builds and it seems to have solved the issue.

Update .babelrc to cover production builds.
Closes https://github.com/facebook/relay/pull/1015

Reviewed By: steveluscher

Differential Revision: D3143803

Pulled By: kassens

fb-gh-sync-id: 64d67cf7dd59c4a8365f29dd2317d71ea2a0c6be
fbshipit-source-id: 64d67cf7dd59c4a8365f29dd2317d71ea2a0c6be