Add deprecation pathway for mutation config fields
Summary:This should be a behavioral no-op, but I'm preparing for the ability to mark mutation config fields as deprecated. I tested this manually by temporarily marking one of the existing fields as deprecated (`s/REQUIRED/DEPRECATED/`) and then using an assertion like this:

```
it('warns about deprecated keys', () => {
  validateMutationConfig({
    ...config,
    connectionName: 'todos',
  }, 'MyMutation');
  expect([
    'validateMutationConfig: `%s` config on `%s` %s `%s`.',
    'NODE_DELETE',
    'MyMutation',
    'has deprecated property',
    'connectionName'
  ]).toBeWarnedNTimes(1);
});
```

Once we have a mutation config that actually has a deprecated attribute then I can commit a test, and it will look pretty much like that.

Making the deprecation message customizable would be a trivial extension of this (for example, if we rename a key, which is actually what led me to start looking at this, we can point to the new name for the key in the deprecation message).
Closes https://github.com/facebook/relay/pull/1029

Reviewed By: steveluscher

Differential Revision: D3154222

Pulled By: wincent

fb-gh-sync-id: 5d2254b4feeeff960fb68ce4a0bbecff8477177f
fbshipit-source-id: 5d2254b4feeeff960fb68ce4a0bbecff8477177f