Remove an unnecessary guard
Summary:In code review for D3136035 yungsters suggested hoisting a guard up to the callsites. Turns out the callsites already have guards (either due to prior `invariant` calls or an explicit `if`), so the guard isn't needed at all (in fact, was in there as a remnant of an intermediate state I had while preparing that diff).

So, let's just remove the guard.
Closes https://github.com/facebook/relay/pull/1022

Reviewed By: kassens

Differential Revision: D3145369

Pulled By: wincent

fb-gh-sync-id: b607170f7b53fc687abad56b46a54896630074a4
fbshipit-source-id: b607170f7b53fc687abad56b46a54896630074a4