Don't apply optimistic update as a side-effect of creating a transaction
Summary:The primary goal of this commit is to "purify" transaction creation by removing the side-effect which applies the optimistic update, and instead making the effect explicit inside `RelayEnvironment`.

There are a few small things that I'm doing here; I'm open to splitting this up into separate diffs if people want it:

- Move side-effect from `RelayMutationQueue` into `RelayEnvironment`.
- Make the `applyOptimistic()` and `commit()` methods in `RelayMutationTransaction` part of a "fluent" interface, so that method calls can be conveniently chained elsewhere (in `RelayEnvironment`).
- Destructure `RelayMutationTransactionStatus` constants in `RelayMutationTransaction` and the `RelayMutationQueue` tests to make things a little more terse and readable.
- Swap `jest.genMockFunction` and `jest.genMockFunction().mockImplementation` for `jest.fn`.

And of course, update the tests so that things still pass.

Depends on D3161029.

Reviewed By: josephsavona, kassens

Differential Revision: D3161900

fb-gh-sync-id: 5fd54bbbeb92e0bd547a2328da2684c938e80fac
fbshipit-source-id: 5fd54bbbeb92e0bd547a2328da2684c938e80fac