Replace toEqualIgnoringMetadata with toMatchRecord
Summary:As planned in D3183382, I'm removing `toEqualIgnoringMetadata` here and swapping in `matchRecord`, but I am keeping the `toEqualIgnoringMetadata` tests because they add value.

Running the tests revealed a little "bug" in `matchRecord`, which was that it would declare `1` and `null` to be equal to one another. I've fixed that in a minimal way by adding a `typeof` check. I've also made the failure output a little more readable:

Before:

```
Expected  to be bar, but it was baz.
Expected this.that.other to be foo, but it was fizz.
```

After:

```
Expected value to be "bar, but it was "baz".
Expected property at path `this.that.other` to be "foo", but it was "fizz".
```

Bonus:

- Removed `toMatchResult`, which had no callers.
- Made things a bit more general, so we handle values of different types a bit better.

Depends on D3183400.

Reviewed By: josephsavona

Differential Revision: D3187595

fb-gh-sync-id: cff0c73c42d34b669cb106a7d770d1892f2c6947
fbshipit-source-id: cff0c73c42d34b669cb106a7d770d1892f2c6947