Fix OSS-only RelayGraphQLMutation CI failure
Summary:This test is consistently failing in Travis CI (and in a clean OSS local checkout) and never failing in our internal repo.

Force whatever internal state is at play to be cleared with a call to `jest.resetModuleRegistry()`, and now we're happy in both places.

?\_(?)_/?

Closes: https://github.com/facebook/relay/issues/1065

Example build failure: https://travis-ci.org/facebook/relay/jobs/124261109

Reviewed By: josephsavona

Differential Revision: D3197446

fb-gh-sync-id: bcd42e1cb2ccb7dfe290a46d42e26de7209130cd
fbshipit-source-id: bcd42e1cb2ccb7dfe290a46d42e26de7209130cd