Make RelayGraphQLMutation method naming more consistent
Summary:Before the addition of `RelayGraphQLMutation`, we had the following naming patterns for applying and committing mutations:

- `RelayMutationTransaction`
  - `commit`
  - `applyOptimistic`
  - `recommit`
  - `rollback`
- `RelayMutationQueue`
  - `commit`
  - `applyOptimistic`
  - `rollback`
- `RelayEnvironment`
  - `applyUpdate`
  - `commitUpdate` (was `update`)

So, evidently, we had some inconsistency between our public API (`RelayEnvironment`) and our internal implementation classes. I made things worse when I added `RelayGraphQLMutation`, because I added a third variant.

- `RelayGraphQLMutation`
  - `commitUpdate`
  - `applyUpdate`

I did this because I wanted it to be easy to grep for the `RelayGraphQLMutation` methods, but on looking back at this I think that was a mistake. Let's unify it with the "internal" classes so that they all share this naming:

- `RelayGraphQLMutation`
  - `commitUpdate` -> `commit`
  - `applyUpdate` -> `applyOptimistic`

Reviewed By: yuzhi

Differential Revision: D3208389

fb-gh-sync-id: 0c54cb1fc02eb59d5fd9e949959702ecd6e06616
fbshipit-source-id: 0c54cb1fc02eb59d5fd9e949959702ecd6e06616