Added basic docs for Relay.Renderer
Summary:
It is public API but has not documentation whatsoever, which can be confusing as this could indicate that the RelayRenderer API is not (yet) meant to be used. This should make it easier to discover and more clear that the RelayRenderer API can be used.

Specifically the usage of `RelayRenderer` would solve some of our use-cases. Also having some documentation for this would mean that https://github.com/relay-tools/react-router-relay/pull/123 could be merged in.

Happy to add any changes etc. I just thought it would be better to have some documentation, rather than none.
Closes https://github.com/facebook/relay/pull/1076

Reviewed By: josephsavona

Differential Revision: D3219946

Pulled By: wincent

fb-gh-sync-id: 616d2763c5f8d2a0d84af7a12ec77da51c161b21
fbshipit-source-id: 616d2763c5f8d2a0d84af7a12ec77da51c161b21