Stop relying on clientRequestId in graphql response.
It is easy to keep track of this id internally without sending it over the
network.  However, requests continue to pass the id as input for requests,
since servers might have defined the input field as non-null. The mutation
spec has been updated to say that servers should now make the field nullable.