Allow disabling Relay container shouldComponentUpdate
Summary:
The core idea here makes sense - providing a way to override the default `shouldComponentUpdate` to return a different value. However, the name `pure` is simply confusing. The intention was that `pure: false` acted as `shouldComponentUpdate: () => true` - but `pure` also suggests that a container is a pure function, such that `shouldComponentUpdate: () => false` if the props are unchanged (which *is* the default already).

Instead, a container definition can now optionally include a `shouldComponentUpdate: () => boolean` function. If this is specified, this function *always* overrides the default implementation (i.e. there is no fall-through to the default). To ensure a re-render when data is passed on context, use `shouldComponentUpdate: () => true`.

This PR introduces an optional flag `pure` to `RelayContainer` which overrides `shouldComponentUpdate` to return true.

Closes https://github.com/facebook/relay/issues/674.
Closes https://github.com/facebook/relay/pull/897

Reviewed By: kassens, wincent

Differential Revision: D3179446

Pulled By: josephsavona

fb-gh-sync-id: d379c23685bf48bb11f1e2d9b39abc98be965644
fbshipit-source-id: d379c23685bf48bb11f1e2d9b39abc98be965644