RelayReadyStateRenderer: re-render when root ids change
Summary:
`RelayReadyStateRenderer` currently constructs the root fragment pointers once the first time that the state is `ready`, and then only updates them if the container or queryconfig changes.

This caused a bug when disk cache was enabled as follows:
- Relay reads the cached root call id for a query, constructs a fragment pointer accordingly, renders the root container with this root id.
- Fresh server results are received and written to the store, but the fragment pointer is not updated
- Container still renders the cached id
- Container calls `forceFetch` which refetches via the old, cached id

The fix here is to only memoize the `querySet`, not the props, which are now always reconstructed. This also fixes an issue where a root fragment pointer could have been null due to the whole query being deferred.

Reviewed By: yuzhi

Differential Revision: D3247996

fb-gh-sync-id: f3233e84dc2ddcadab28c1b9141a94372ce6e08d
fbshipit-source-id: f3233e84dc2ddcadab28c1b9141a94372ce6e08d