Fix issue with synchronous mutation rollback + apply
Summary:
Fixes https://github.com/facebook/relay/issues/1081 and an internal issue with
synchronously rolling back an optimistic-only mutation and re-applying another
with overlapping data. The bug is as follows:

- Mutation 1 applied: record A is marked as affected by that mutation.
- (later) Mutation 1 rolled back: the mutation is removed from the mutation
  queue, and the queuedData change is scheduled for removal with
  `resolveImmediagte`
- In the same event loop, mutation 2 is applied also affecting record A.
- Affected components re-read data for record A and attempt to read the status
  of optimistic mutations that affect it. Record A still lists mutation 1 & 2,
  but 1 has been removed from the queue. Reading its status breaks the invariant
  in `getTransaction`.

Making `applyUpdate` also async would make apps less responsive - changes would
not appear in the same frame. Making `rollback` synchronous could be very slow.
The fix is to allow `getTransaction` to return a nullable transaction so that
the reader continues to work. A new integration test makes this case clear and
ensures it will continue to work.

Reviewed By: wincent

Differential Revision: D3270965

fb-gh-sync-id: 0257a11b8aebf31f2cd7b642c700c7e3e74cb697
fbshipit-source-id: 0257a11b8aebf31f2cd7b642c700c7e3e74cb697