Use alias when constructing serialization key
Summary:
Relay can currently construct invalid GraphQL in the case where two sibling fields end up with the same calls & values. For example in the following, Relay would produce the same serialization key (alias) if $env = WEB:

```graphql
fragment on Foo {
  defaultSettings {
    web: notifications(environment: WEB)
    foo: notifications(environment: $env)
  }
}
```

This PR ensures that unique aliases are produced in this case by using both the canonical schema name *and* the alias (if defined) when constructing the alias (in addition to calls & values).
Closes https://github.com/facebook/relay/pull/1129

Reviewed By: kassens

Differential Revision: D3274879

Pulled By: josephsavona

fbshipit-source-id: 3718e26ac132ec6bc2d9c631f0e9cf27dfe870e0