Make props.relay.variables the prepared variables
Summary:
Changes `props.relay.variables` to reflect the output of `prepareVariables`.
This is a long-standing feature request; the fact that variables were the
un-prepared values is surprising/unintuitive and can lead to bugs (since the
variables are different than what is used to actually fetch the data).

The change is in `RelayContainer#_initialize`, where the merged initial/parent
variables are now run through `prepareVariables` if specified.

Fixes #578.

Reviewed By: kassens

Differential Revision: D3271810

fbshipit-source-id: 4d7371f5d5673f5ac462296377203de395e2bd6c