Teach `RelayGraphQLMutation` to pass through `RelayMutationConfig` objects.
Summary:
Because `RelayGraphQLMutation` ends up enqueuing a transaction in the `RelayMutationQueue`, which is primarily designed to handle transactions related to `RelayMutation` instances, there is a point at which the mutation has been fired off and there's not a lot that the low-level API can do to influence its status. It's in the queue and other parts of Relay are now responsible for handling it.

As such, if we want the store to be properly updated for the more "magical" mutation types such as `RANGE_ADD` then we need to pass along a mutation config. I expect this to change in the future, but for now this is the easiest way to make it happen.

I thought about a few ways to do this, such as passing the configs in at construction time, adding setters etc. In the end I went with passing in the config at apply/commit time because I wanted to make the smallest possible change to the API, especially given that I want this to be a temporary thing. Putting the configs in at this level makes them less prominent, which is exactly how I want them to be so as to maximize the room for changing how all this works in the future.

Depends on D3278912.

Reviewed By: josephsavona

Differential Revision: D3279015

fbshipit-source-id: 37632eb9d7105548fd51151b684fee8452c5c434