Fix formatErrors when offset is negative
Summary:
I was looking at this stackoverflow question: https://stackoverflow.com/questions/36948277/relay-js-error-repeat-argument-must-be-greater-than-or-equal-to-0-and-not-be

It seems like if your server returns errors with column `0`, offset would be `-1`, and `repeat` would 💥  because `.repeat` takes a count argument which is
> An integer between 0 and +∞.

This fix is just `Math.max(0, offset)`.

I was looking for the spec for errors, I'm not totally sure if `0` is an accepted value for column in the first place so I might be missing context.
Closes https://github.com/facebook/relay/pull/1105

Reviewed By: josephsavona

Differential Revision: D3284541

Pulled By: kassens

fbshipit-source-id: 68be2183e453e6f3559015b65ad07bd386312033