RelayContainer: call prepareVariables with raw values
Summary:
D3271810 changed `RelayContainer` to set props.relay.variables to the prepared variables. As a result of this change `setVariables` was using the correct,  prepared variables, but the variables were not correctly updated when the fetch completed - in some cases the variables could be prepared twice, which would break components with non-idempotent `prepareVariables` functions.

This changes RelayContainer to call `prepareVariables` as follows:
- In componentWillMount/componentWillReceiveProps: when new (unprepared) vars are received they are prepared. The original variables are stored as `this.state.rawVariables` and prepared variables are stored as `this.state.relayProp.variables`.
- In setVariables: the previous & next raw variables are merged and then prepared, with the prepared vars used both to fetch the data. Upon completion of the fetch, `relayProp.variables` is updated with the new prepared variables and `state.rawVariables` is updated with the new unprepared variables.

Reviewed By: yuzhi

Differential Revision: D3285585

fbshipit-source-id: 22baae4bb0769dc63fc22305e3a9f7b49882b115